+++ cmd/process-agent/command/main_common.go
@@ -19,8 +19,8 @@ import (
 	"github.com/DataDog/datadog-agent/comp/api/authtoken/fetchonlyimpl"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/comp/core/configsync/configsyncimpl"
 	configsync "github.com/DataDog/datadog-agent/comp/core/configsync/def"
+	configsyncimpl "github.com/DataDog/datadog-agent/comp/core/configsync/impl"
 	logcomp "github.com/DataDog/datadog-agent/comp/core/log/def"
 	"github.com/DataDog/datadog-agent/comp/core/pid"
 	"github.com/DataDog/datadog-agent/comp/core/pid/pidimpl"
