+++ cmd/security-agent/main_windows.go
@@ -28,8 +28,8 @@ import (
 	"github.com/DataDog/datadog-agent/comp/api/authtoken/fetchonlyimpl"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/comp/core/configsync/configsyncimpl"
 	configsync "github.com/DataDog/datadog-agent/comp/core/configsync/def"
+	configsyncimpl "github.com/DataDog/datadog-agent/comp/core/configsync/impl"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	"github.com/DataDog/datadog-agent/comp/core/secrets"
 	"github.com/DataDog/datadog-agent/comp/core/settings"
