+++ comp/core/configsync/impl/module_integration_test.go
