+++ comp/core/configsync/impl/module_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-package configsyncimpl
+package impl
 
 import (
 	"testing"
