+++ comp/core/configsync/impl/sync_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package configsyncimpl
+package impl
 
 import (
 	"context"
