+++ comp/core/configsync/impl/test_common.go
@@ -5,7 +5,7 @@
 
 //go:build test
 
-package configsyncimpl
+package impl
 
 import (
 	"context"
