+++ modules.yml
@@ -17,6 +17,8 @@ modules:
   comp/api/authtoken: default
   comp/core/config:
     used_by_otel: true
+  comp/core/configsync/def:
+    used_by_otel: true
   comp/core/flare/builder:
     used_by_otel: true
   comp/core/flare/types:
